Move ConfigKeyUtils to common module
